refactor(normalizer): move atoms to a separate file
